Decoupled the UI and the reading thread.
More specifically, I made it so that the ROV reading code isn't dependent
on the user interface thread. This should lead to a nicer organization of
the code, especially since the UI will touching almost every part of the
codebase, and not every part of the codebase should know about every other
part.